+++ gazeplay/src/main/java/net/gazeplay/utils/Home.java
@@ -1,4 +1,4 @@
-package utils.games;
+package net.gazeplay.utils;
 
 import javafx.scene.image.Image;
 import javafx.scene.paint.ImagePattern;
